vector-db-agents/requirements.txt
qdrant-client==1.6.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3